% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_grpFabs.r
\name{GFabs}
\alias{GFabs}
\title{A Group Forward and Backward Stagewise (GFabs) algorithm for smoothed partial rank loss function (SPR) with the Group Lasso penalty.}
\usage{
GFabs(
  W,
  y,
  group,
  status = NULL,
  sigma = NULL,
  weight = NULL,
  model = "spr",
  type = "L2",
  back = TRUE,
  stoping = TRUE,
  eps = 0.01,
  xi = 10^-6,
  iter = 10^4,
  lambda.min = NULL
)
}
\arguments{
\item{W}{The design matrix.}

\item{y}{The survival outcome.}

\item{group}{The grouping vector.}

\item{status}{The censoring indicator.}

\item{sigma}{The smoothing parameter in SPR.}

\item{weight}{The weight vector of groups.}

\item{model}{The loss function used.}

\item{type}{The norm of penalty used.}

\item{back}{The indicator of whether to take backward steps.}

\item{stoping}{The indicator of whether to stop iteration when lambda is less than lambda.min.}

\item{eps}{The step size for GFabs.}

\item{xi}{The threshhold for GFabs.}

\item{iter}{The maximum number of outer-loop iterations allowed.}

\item{lambda.min}{The smallest value for lambda, as a fraction of lambda.max.}
}
\value{
A list.
\itemize{
  \item beta - estimation of covariates
  \item lambda - lambda sequence
  \item direction - direction of GFabs
  \item active - active sets
  \item iter - iterations
  \item BIC - bic criteria
  \item group - The grouping vector.
  \item opt - position of the optimal tuning based on BIC.
}
}
\description{
A Group Forward and Backward Stagewise (GFabs) algorithm for smoothed partial rank loss function (SPR) with the Group Lasso penalty.
}
\examples{
W = matrix(rnorm(80), 20, 4)
b = c(5, 5, -5, -5)
e = c(0.7*rnorm(20)+0.3*rcauchy(20))
y = W \%*\% b + e
group <- c(1, 1, 2, 2)
fit <- GFabs(W, y, group)
}
